feat: add useStopWatchHook
